refactor(ser,de): move both body deserializer and serializer into a new module each
