+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![allow(incomplete_features)]
 #![feature(return_position_impl_trait_in_trait)]
+pub mod deserializer;
 pub mod error;
 pub mod handle_selector;
 pub mod handler;
@@ -10,4 +11,5 @@ pub mod result;
 pub mod router;
 pub mod runner_input;
 pub mod runner_output;
+pub mod serializer;
 pub mod server;
